chore(#49):  Split DB service file
Split the DB service file into individual service and scope files.
