+++ workspaces/backend/gql/src/scopes/index.ts
@@ -1,0 +1,3 @@
+export * from "./withQueryRunner.js";
+export * from "./withQueryRunnerTransaction.js";
+export * from "./withTransaction.js";
