+++ workspaces/backend/gql/src/services/EntityManagerService.ts
@@ -1,0 +1,7 @@
+import { ServiceName } from "../enums/service.js";
+import { Context } from "effect";
+import type { EntityManager } from "typeorm";
+
+export class EntityManagerService extends Context.Tag(
+  ServiceName.EntityManager,
+)<EntityManagerService, EntityManager>() {}
