+++ workspaces/backend/gql/src/services/QueryRunnerService.ts
@@ -1,0 +1,8 @@
+import { ServiceName } from "../enums/service.js";
+import { Context } from "effect";
+import type { QueryRunner } from "typeorm";
+
+export class QueryRunnerService extends Context.Tag(ServiceName.QueryRunner)<
+  QueryRunnerService,
+  QueryRunner
+>() {}
