+++ workspaces/backend/gql/src/services/index.ts
@@ -1,1 +1,3 @@
-export * from "./db.js";
+export * from "./DataSourceService.js";
+export * from "./EntityManagerService.js";
+export * from "./QueryRunnerService.js";
